cleaned up code, commented
