+++ client/src/components/page-components/Headlines.tsx
@@ -1,13 +1,24 @@
 import HeadlineCard from "./HeadlineCard"
 
-const Headlines = () => {
-return <div className="w-full flex justify-between px-10">
-    <HeadlineCard headline={"This is an eye catching headline"}/>
-    <HeadlineCard headline={"This is another great eye catching headline"}/>
-    <HeadlineCard headline={"This is yet another great eye catching headline"}/>
-    <HeadlineCard headline={"This is the final eye catching headline"}/>
-</div>
+type Props = {
+    stories: Story[]
+}
 
+type Story = {
+    name: string,
+    author: string,
+    body: string,
+    source_url: string,
+    cover_url: string,
+    category: string
+}
+
+const Headlines = ( props: Props ) => {
+    return (
+        <div className="w-full flex justify-between px-10">
+            {props.stories?.map(story => <HeadlineCard img={story.cover_url} headline={story.name}/>)}
+        </div>
+    )
 }
 
 export default Headlines
